Search/ES: remove the dependency to takari-cpsuite (it's not used)
The dependency was probably added when copying dependencies from
hibernate-search-elasticsearch, where we need it to have an easy way to
execute a set of tests based on a path regexp.
Maybe we had to do this because we used to import test-jars (for which
transitive dependencies do not work).

But here in our test case template, it's probably overkill: we will
never have many tests.